Card editing uses templates and add discard button
Card editing used to be done by changing the HTML of a card using
jQuery which was undesireable so now there is a session variable "edit"
which contains 0 for no card being edited else the mongo id of the card
being edited. This session variable determines which template is
rendered by the card_wrapper. The discard button simple resets the
session variable without saving the changes.