Change default card view to be more compact.
I have limited the length of the description displayed in the view as
well as adding a progress bar for subtasks and reducing the size of the
assigness list.  The aim of this patch is to reduce the size of the
cards vertically with the intention of adding more extended view that
the user can access on click in a further patchset.